Add error handling and authorisation to controllers

+++ OrderingSystemProject/Controllers/MenuItemController.cs
@@ -23,18 +23,36 @@ public class MenuItemController : Controller
             return RedirectToAction("Login", "Employees");
         }
 
-        List<MenuItem> menuItems = _menuItemService.GetAll();
-        var menuManagementViewMode = new MenuManagementViewModel(
-            menuItems,
-            CardFilterType.ALL,
-            CategoryFilterType.ALL
-        );
-        return View(menuManagementViewMode);
+        try
+        {
+            List<MenuItem> menuItems = _menuItemService.GetAll();
+            var menuManagementViewModel = new MenuManagementViewModel(
+                menuItems,
+                CardFilterType.ALL,
+                CategoryFilterType.ALL
+            );
+            return View(menuManagementViewModel);
+        }
+        catch (Exception e)
+        {
+            TempData["Exception"] = $"Exception occured: {e.Message}";
+            var menuManagementViewModel = new MenuManagementViewModel(
+                new List<MenuItem>(),
+                CardFilterType.ALL,
+                CategoryFilterType.ALL
+            );
+            return View(menuManagementViewModel);
+        }
     }
 
     [HttpGet]
     public IActionResult Create()
     {
+        if (IsUnauthorisedUser())
+        {
+            return RedirectToAction("Login", "Employees");
+        }
+
         return View();
     }
 
@@ -57,18 +75,36 @@ public class MenuItemController : Controller
     [HttpGet("MenuItem/Delete/{menuItemId}")]
     public ActionResult Delete(int? menuItemId)
     {
+        if (IsUnauthorisedUser())
+        {
+            return RedirectToAction("Login", "Employees");
+        }
+
         if (menuItemId == null)
         {
             return NotFound();
         }
 
-        MenuItem? menuItem = _menuItemService.GetById((int)menuItemId);
-        return View(menuItem);
+        try
+        {
+            MenuItem? menuItem = _menuItemService.GetById((int)menuItemId);
+            return View(menuItem);
+        }
+        catch (Exception e)
+        {
+            TempData["Exception"] = $"Exception occured: {e.Message}";
+            return RedirectToAction(nameof(Index));
+        }
     }
 
     [HttpPost]
     public ActionResult Delete(MenuItem menuItem)
     {
+        if (IsUnauthorisedUser())
+        {
+            return RedirectToAction("Login", "Employees");
+        }
+
         try
         {
             _menuItemService.Delete(menuItem);
@@ -85,17 +121,36 @@ public class MenuItemController : Controller
     [HttpGet("MenuItem/Edit/{menuItemId}")]
     public IActionResult Edit(int? menuItemId)
     {
+        if (IsUnauthorisedUser())
+        {
+            return RedirectToAction("Login", "Employees");
+        }
+
         if (menuItemId == null)
         {
             return NotFound();
         }
-        var menuItem = _menuItemService.GetById((int)menuItemId);
-        return View(menuItem);
+
+        try
+        {
+            var menuItem = _menuItemService.GetById((int)menuItemId);
+            return View(menuItem);
+        }
+        catch (Exception e)
+        {
+            TempData["Exception"] = $"Exception occured: {e.Message}";
+            return RedirectToAction(nameof(Index));
+        }
     }
 
     [HttpPost]
     public IActionResult Edit(MenuItem menuItem)
     {
+        if (IsUnauthorisedUser())
+        {
+            return RedirectToAction("Login", "Employees");
+        }
+
         try
         {
             _menuItemService.Update(menuItem);
@@ -111,15 +166,20 @@ public class MenuItemController : Controller
 
     public IActionResult Filter(CardFilterType cardFilterType, CategoryFilterType categoryFilterType)
     {
+        if (IsUnauthorisedUser())
+        {
+            return RedirectToAction("Login", "Employees");
+        }
+
         try
         {
             List<MenuItem> menuItems = _menuItemService.Filter(categoryFilterType, cardFilterType);
-            var menuManagementViewMode = new MenuManagementViewModel(
+            var menuManagementViewModel = new MenuManagementViewModel(
                 menuItems,
                 cardFilterType,
                 categoryFilterType
             );
-            return View(nameof(Index), menuManagementViewMode);
+            return View(nameof(Index), menuManagementViewModel);
         }
         catch (Exception e)
         {
@@ -131,6 +191,11 @@ public class MenuItemController : Controller
     [HttpPost]
     public IActionResult Activate(int menuItemId)
     {
+        if (IsUnauthorisedUser())
+        {
+            return RedirectToAction("Login", "Employees");
+        }
+
         try
         {
             _menuItemService.Activate(menuItemId);
@@ -147,6 +212,11 @@ public class MenuItemController : Controller
     [HttpPost]
     public IActionResult Deactivate(int menuItemId)
     {
+        if (IsUnauthorisedUser())
+        {
+            return RedirectToAction("Login", "Employees");
+        }
+
         try
         {
             _menuItemService.Deactivate(menuItemId);
